requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
supabase==2.5.1
psycopg2-binary==2.9.9
pillow==10.4.0
django-extensions==3.2.3
pydot==3.0.1
pyparsing==3.1.2
graphviz==0.20.3